+++ earth-commons/src/main/java/com/ani/commons/earth/dto/organization/node/OTNodeAddDto.java
@@ -11,12 +11,12 @@ import java.io.Serializable;
 public class OTNodeAddDto extends AniDto implements Serializable {
 
     private static final long serialVersionUID = -463500843397755139L;
-    private Long orgId;
+    private Integer orgId;
     private String name;
     private NodeType type;
     private OTNodeOpDto otNodeOpDto;
 
-    public OTNodeAddDto(Long orgId, String name, NodeType type, OTNodeOpDto otNodeOpDto) {
+    public OTNodeAddDto(Integer orgId, String name, NodeType type, OTNodeOpDto otNodeOpDto) {
         this.orgId = orgId;
         this.name = name;
         this.type = type;
@@ -26,11 +26,11 @@ public class OTNodeAddDto extends AniDto implements Serializable {
     public OTNodeAddDto() {
     }
 
-    public Long getOrgId() {
+    public Integer getOrgId() {
         return orgId;
     }
 
-    public void setOrgId(Long orgId) {
+    public void setOrgId(Integer orgId) {
         this.orgId = orgId;
     }
